requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
slowapi==0.1.9
python-multipart==0.0.12
alembic==1.13.2
pydantic[email]